Handle errors consistently.
