+++ handlers/retrieve_resource_test.go
@@ -13,7 +13,7 @@ func TestRetrieveHappyPath(t *testing.T) {
 	r := gofight.New()
 	repo := NewMockDatabase(&datautils.Resource{})
 	r.GET("/v1/resource/99").
-		Run(handler(repo, NewMockStream(""), nil),
+		Run(handler(repo, nil, nil),
 			func(r gofight.HTTPResponse, rq gofight.HTTPRequest) {
 				assert.Equal(t, http.StatusOK, r.Code)
 			})
@@ -22,18 +22,19 @@ func TestRetrieveHappyPath(t *testing.T) {
 func TestRetrieveNotFound(t *testing.T) {
 	r := gofight.New()
 	r.GET("/v1/resource/100").
-		Run(handler(NewMockDatabase(nil), NewMockStream(""), nil),
+		Run(handler(NewMockDatabase(nil), nil, nil),
 			func(r gofight.HTTPResponse, rq gofight.HTTPRequest) {
 				assert.Equal(t, http.StatusNotFound, r.Code)
 			})
 }
 
-// TODO: Error handling
-// func TestRetrieveError(t *testing.T) {
-// 	r := gofight.New()
-// 	r.GET("/v1/resource/100").
-// 		Run(setupFakeRuntime(mockErrorRepo()),
-// 			func(r gofight.HTTPResponse, rq gofight.HTTPRequest) {
-// 				assert.Equal(t, http.StatusUnprocessableEntity, r.Code)
-// 			})
-// }
+func TestRetrieveError(t *testing.T) {
+	r := gofight.New()
+	assert.Panics(t,
+		func() {
+			r.GET("/v1/resource/100").
+				Run(handler(NewMockErrorDatabase(), nil, nil),
+					func(r gofight.HTTPResponse, rq gofight.HTTPRequest) {})
+
+		})
+}
