+++ handlers/stub_database_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/sul-dlss-labs/taco/db"
 	"github.com/sul-dlss-labs/taco/storage"
 	"github.com/sul-dlss-labs/taco/streaming"
-	"github.com/sul-dlss-labs/taco/uploaded"
 )
 
 func handler(database db.Database, stream streaming.Stream, storage storage.Storage) http.Handler {
@@ -33,22 +32,6 @@ func NewMockDatabase(record *datautils.Resource) db.Database {
 	return &MockDatabase{CreatedResources: []datautils.Resource{}, record: record}
 }
 
-type MockStream struct {
-	message string
-}
-
-func NewMockStream(message string) streaming.Stream {
-	return &MockStream{message: message}
-}
-
-type MockStorage struct {
-	CreatedFiles []*uploaded.File
-}
-
-func NewMockStorage() storage.Storage {
-	return &MockStorage{CreatedFiles: []*uploaded.File{}}
-}
-
 func (d *MockDatabase) Insert(params datautils.Resource) error {
 	d.CreatedResources = append(d.CreatedResources, params)
 	return nil
@@ -63,16 +46,6 @@ func (d *MockDatabase) Read(id string) (*datautils.Resource, error) {
 
 func (d *MockDatabase) Update(params datautils.Resource) error {
 	return nil
-}
-
-func (s *MockStream) SendMessage(message string) error {
-	return nil
-}
-
-func (s *MockStorage) UploadFile(id string, file *uploaded.File) (*string, error) {
-	s.CreatedFiles = append(s.CreatedFiles, file)
-	path := "s3FileLocation"
-	return &path, nil
 }
 
 type MockErrorDatabase struct {
@@ -91,15 +64,5 @@ func (d *MockErrorDatabase) Update(params datautils.Resource) error {
 }
 
 func (d *MockErrorDatabase) Read(id string) (*datautils.Resource, error) {
-	return nil, errors.New("Broken")
-}
-
-type MockErrorStorage struct{}
-
-func NewMockErrorStorage() storage.Storage {
-	return &MockErrorStorage{}
-}
-
-func (d *MockErrorStorage) UploadFile(id string, file *uploaded.File) (*string, error) {
 	return nil, errors.New("Broken")
 }
